using httpproxy library as 3rd party dependency
