--- x/local-proxy/proxy/httpproxy/AUTHORS
@@ -1,11 +1,0 @@
-# This is the official list of authors for copyright purposes.
-# This file is distinct from the CONTRIBUTORS files.
-
-# Names should be added to this file as one of
-#     Organization's name
-#     Individual's name <submission email address>
-#     Individual's name <submission email address> <email2> <emailN>
-
-# Please keep the list sorted.
-
-Orkun Karaduman <orkunkaraduman@gmail.com>
