--- x/local-proxy/proxy/httpproxy/CONTRIBUTORS
@@ -1,14 +1,0 @@
-# This is the official list of people who can contribute
-# (and typically have contributed) code to this repository.
-# The AUTHORS file lists the copyright holders; this file
-# lists people.
-
-# Names should be added to this file like so:
-#     Individual's name <submission email address>
-#     Individual's name <submission email address> <email2> <emailN>
-
-# Please keep the list sorted.
-
-Orkun Karaduman <orkunkaraduman@gmail.com>
-Yasin Özel <me@yozel.co>
-Pavel Griaznov <grbitt@gmail.com>
